+++ packages/framework/test/di_test.dart
@@ -27,7 +27,7 @@ void main() {
     // Inject some todos
     app.container!.registerSingleton(Todo(text: TEXT, over: OVER));
     app.container!.registerFactory<Future<Foo>>((container) async {
-      var req = container.make<RequestContext>()!;
+      var req = container.make<RequestContext>();
       var text = await utf8.decoder.bind(req.body!).join();
       return Foo(text);
     });