+++ packages/container/angel_container/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 3.1.1
 
 * Updated `_ReflectedMethodMirror` to have optional `returnType` parameter
-* Updated all the methods of `Container` class to return non nullable type
+* Updated `Container` to handle non nullable type
 
 ## 3.1.0