+++ packages/container/angel_container/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 6.0.0
+
+* Updated to min SDK 2.15.x
+* Removed `error`
+
+## 5.0.0
+
+* Skipped release
+
+## 4.0.0
+
+* Skipped release
+  
 ## 3.1.1
 
 * Updated `_ReflectedMethodMirror` to have optional `returnType` parameter